requirements.txt
asgiref==3.9.0
azure-core==1.35.1
azure-storage-blob==12.26.0
certifi==2025.6.15
cffi==2.0.0
charset-normalizer==3.4.2
coverage==7.13.4
cryptography==46.0.1
dj-database-url==3.0.1
Django==5.2.4
django-allauth==65.14.1
django-filter==25.2
django-formtools==2.5.1
django-storages==1.14.6
djangorestframework==3.17.0
docker==7.1.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.3.0
isodate==0.7.2
Markdown==3.10.2
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg==3.2.10
pycparser==2.23
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.12.0
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.4
rich==14.0.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.9.0
